Move state mod to combinator and some changes
