+++ src/interpolate.rs
@@ -77,147 +77,6 @@ use bevy::prelude::*;
 use crate::utils::color_lerp;
 use crate::{tween, BevyTweenRegisterSystems};
 
-mod state {
-    use tween::{ComponentTween, Tween};
-
-    use super::*;
-
-    // pub trait CombinatorState<S> {
-    //     type Out<FO>;
-    //     fn state_for<F, FO>(&mut self, f: F) -> Self::Out<FO>
-    //     where
-    //         F: FnOnce(&mut S) -> FO;
-    // }
-
-    // impl<T> CombinatorState<T> for T {
-    //     type Out<FO> = FO;
-
-    //     fn state_for<F, FO>(&mut self, f: F) -> Self::Out<FO>
-    //     where
-    //         F: FnOnce(&mut T) -> FO,
-    //     {
-    //         f(self)
-    //     }
-    // }
-
-    /// Constructor for [`GenericTweenState`]
-    pub fn tween_state<T, V>(target: T, value: V) -> GenericTweenState<T, V> {
-        GenericTweenState { target, value }
-    }
-
-    pub struct GenericTweenState<T, V> {
-        target: T,
-        value: V,
-    }
-
-    impl<T, V> GenericTweenState<T, V> {
-        pub fn new(target: T, value: V) -> GenericTweenState<T, V> {
-            GenericTweenState { target, value }
-        }
-    }
-
-    impl<T, V> GenericTweenState<T, V>
-    where
-        T: Clone,
-    {
-        pub fn tween<I>(&mut self, f: impl FnOnce(&mut V) -> I) -> Tween<T, I> {
-            let interpolator = f(&mut self.value);
-            Tween {
-                target: self.target.clone(),
-                interpolator,
-            }
-        }
-    }
-
-    /// Constructor for [`TransformTweenState`]
-    pub fn transform_tween_state(
-        target: tween::TargetComponent,
-        value: Transform,
-    ) -> TransformTweenState {
-        TransformTweenState::new(target, value)
-    }
-
-    pub struct TransformTweenState {
-        target: tween::TargetComponent,
-        value: Transform,
-    }
-
-    impl TransformTweenState {
-        pub fn new(
-            target: tween::TargetComponent,
-            value: Transform,
-        ) -> TransformTweenState {
-            TransformTweenState { target, value }
-        }
-
-        pub fn tween_transform<I>(
-            &mut self,
-            f: impl FnOnce(&mut Transform) -> I,
-        ) -> Tween<tween::TargetComponent, I> {
-            let interpolator = f(&mut self.value);
-            Tween {
-                target: self.target.clone(),
-                interpolator,
-            }
-        }
-
-        pub fn tween_translation<I>(
-            &mut self,
-            f: impl FnOnce(&mut Vec3) -> I,
-        ) -> Tween<tween::TargetComponent, I> {
-            self.tween_transform(|v| f(&mut v.translation))
-        }
-
-        pub fn tween_rotation<I>(
-            &mut self,
-            f: impl FnOnce(&mut Quat) -> I,
-        ) -> Tween<tween::TargetComponent, I> {
-            self.tween_transform(|v| f(&mut v.rotation))
-        }
-
-        pub fn tween_scale<I>(
-            &mut self,
-            f: impl FnOnce(&mut Vec3) -> I,
-        ) -> Tween<tween::TargetComponent, I> {
-            self.tween_transform(|v| f(&mut v.scale))
-        }
-
-        pub fn translation_to(
-            &mut self,
-            to: Vec3,
-        ) -> ComponentTween<Translation> {
-            self.tween_translation(translation_to(to))
-        }
-
-        pub fn rotation_to(&mut self, to: Quat) -> ComponentTween<Rotation> {
-            self.tween_rotation(rotation_to(to))
-        }
-
-        pub fn scale_to(&mut self, to: Vec3) -> ComponentTween<Scale> {
-            self.tween_scale(scale_to(to))
-        }
-
-        pub fn translation_by(
-            &mut self,
-            by: Vec3,
-        ) -> ComponentTween<Translation> {
-            self.tween_translation(translation_by(by))
-        }
-
-        pub fn rotation_by(&mut self, by: Quat) -> ComponentTween<Rotation> {
-            self.tween_rotation(rotation_by(by))
-        }
-
-        pub fn scale_by(&mut self, by: Vec3) -> ComponentTween<Scale> {
-            self.tween_scale(scale_by(by))
-        }
-    }
-}
-
-pub use state::{
-    transform_tween_state, tween_state, GenericTweenState, TransformTweenState,
-};
-
 /// Alias for an `Interpolator` as a boxed trait object.
 pub type BoxedInterpolator<Item> = Box<dyn Interpolator<Item = Item>>;
